formatting

+++ backup-sizer/src/mongodb.com/size-estimator/components/oplog.go
@@ -13,13 +13,13 @@ import (
 type OplogInfo struct {
 	startTS bson.MongoTimestamp
 	endTS   bson.MongoTimestamp
-	size    int
+	size    int64
 }
 
 type OplogStats struct {
 	StartTS            bson.MongoTimestamp
 	EndTS              bson.MongoTimestamp
-	Size               int
+	Size               int64
 	GbPerDay           float64
 	CompressionRatio   float64
 	CompressedGbPerDay float64
@@ -159,12 +159,12 @@ func getMongodVersion(session *mgo.Session) (string, error) {
 	return result["version"].(string), nil
 }
 
-func getOplogSize(session *mgo.Session) (int, error) {
+func getOplogSize(session *mgo.Session) (int64, error) {
 	result := struct {
-		Capped bool `bson:"capped"`
+		Capped  bool  `bson:"capped"`
 		MaxSize int64 `bson:"maxSize"`
-		Size int64 `bson:"size"`
-	} {}
+		Size    int64 `bson:"size"`
+	}{}
 
 	err := GetOplogCollStats(session, &result)
 	if err != nil {
@@ -175,9 +175,9 @@ func getOplogSize(session *mgo.Session) (int, error) {
 	}
 
 	if result.MaxSize != 0 {
-		return int(result.MaxSize), nil
+		return result.MaxSize, nil
 	}
-	return int(result.Size), nil
+	return result.Size, nil
 }
 
 func getOplogColl(session *mgo.Session) (*mgo.Collection, error) {
